New Valid Targets
You can now use "mounts", "castermount", "passengers", and "casterpassenger" in targeting. Useful if you want to have a spell that heals mounts. These will only work if the entities are stacked.

Also added "mobtarget" as a valid target which will actively target the current mobs target in the spell if it can reach it. Useful for mob casting, useless on players.